# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entity.User:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        $ref: '#/definitions/entity.UserGender'
      id:
        type: string
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      profile_picture_url:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  entity.UserGender:
    enum:
    - Male
    - Female
    type: string
    x-enum-varnames:
    - UserGenderMale
    - UserGenderFemale
  entity.UserMeal:
    properties:
      created_at:
        type: string
      custom_meal_name:
        type: string
      id:
        type: string
      meal_date:
        type: string
      meal_time:
        type: string
      meal_type:
        $ref: '#/definitions/entity.UserMealType'
      notes:
        type: string
      total_calories_consumed:
        type: number
      total_carbs_consumed:
        type: number
      total_fat_consumed:
        type: number
      total_protein_consumed:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entity.UserMealType:
    enum:
    - breakfast
    - lunch
    - dinner
    - snack
    - other
    type: string
    x-enum-varnames:
    - UserMealTypeBreakfast
    - UserMealTypeLunch
    - UserMealTypeDinner
    - UserMealTypeSnack
    - UserMealTypeOther
...
  entity.UserWorkoutSession:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      location:
        type: string
      mood_rating:
        type: integer
      notes:
        type: string
      perceived_exertion_rating:
        type: integer
      plan_id:
        type: string
      scheduled_at:
        type: string
      session_name:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/entity.WorkoutSessionStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entity.WorkoutSessionStatus:
    enum:
    - scheduled
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - WorkoutSessionStatusScheduled
    - WorkoutSessionStatusInProgress
    - WorkoutSessionStatusCompleted
    - WorkoutSessionStatusCancelled
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  v1.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Fitness application API for tracking nutrition, meals, and products
  title: Rebound API
  version: "1.0"
paths:
  /api/users:
    get:
      description: Get a paginated list of users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a user
      tags:
      - users
    get:
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update a user
      tags:
      - users
  /api/users/{id}/password:
    put:
      consumes:
      - application/json
      description: Update a user's password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New password
        in: body
        name: password
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update user password
      tags:
      - users
  /api/users/{id}/verify-email:
    put:
      consumes:
      - application/json
      description: Update a user's email verification status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Verification status
        in: body
        name: verified
        required: true
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update email verification status
      tags:
      - users
...
  /meals/user/{userID}:
    get:
      consumes:
      - application/json
      description: Get all meals for a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserMeal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get user meals
      tags:
      - meals
...
  /workout-sessions/user/{userID}:
    get:
      consumes:
      - application/json
      description: Get all workout sessions for a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserWorkoutSession'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get user workout sessions
      tags:
      - workout-sessions
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"